+++ src/backend/main.server.ts
@@ -1,0 +1,2 @@
+
+export { AppServerModule } from './app.server.module';
