+++ src/frontend/app/app.component.ts
