+++ src/frontend/tsconfig.app.json
@@ -1,12 +1,12 @@
 /* To learn more about this file see: https://angular.io/config/tsconfig. */
 {
-  "extends": "./tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "./out-tsc/app",
     "types": []
   },
   "files": [
-    "src/main.ts"
+    "./main.ts"
   ],
   "include": [
     "src/**/*.d.ts"
