+++ src/frontend/webpack-front.config.js
