--- src/main.server.ts
@@ -1,2 +1,0 @@
-
-export { AppServerModule } from './app/app.server.module';
